// src/types/projectAbout.ts
export interface ProjectAboutData {
  projectTitle: string;
  projectDescription: string;
  projectLink: string;
  projectImage: string;
  projectObjDescription: string;
  projectObjImage: string;
  projectName1: string;
  projectName1Description: string;
  projectName1Image: string;
  projectName2: string;
  projectName2Description: string;
  projectName2Image: string;
  theme: string;
}